Refactored into separate UI components
